project/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
bcrypt
Flask-CORS
marshmallow
